# openapi.yaml
openapi: 3.0.3
info:
  title: Object CRUD
  description: |-
    Sample Object CRUD
  version: 1.0.11
externalDocs:
  description: Link to the repository on Github
  url: http://swagger.io
servers:
  - url: http://localhost:5077/
tags:
  - name: Object
    description: CRUD over Object table
    
    
paths:
  /Object:
    post:
      tags:
        - Object
      summary: Add a new object
      description: Add a new object
      operationId: addObject
      requestBody:
        description: Create a new object
        required: true
        content: 
          application/json:
              schema:
                $ref: '#/components/schemas/Object' 
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Object'   
        '400':
          description: Invalid input
        '422':
          description: Validation exception
            
  /Object/{objectId}:
    get:
      tags:
        - Object
      summary: Find Object by ID
      description: Returns a single object
      operationId: getObjectById
      parameters:
        - name: objectId
          in: path
          description: ID of object to return
          required: true
          schema:
            type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Object'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Object'
        '400':
          description: Invalid ID supplied
        '404':
          description: Object not found
            
    put:
      tags:
        - Object
      summary: Updates an object
      operationId: updateObject
      parameters:
        - name: objectId
          in: path
          description: ID of object that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
        - name: name
          in: query
          description: Name of object that needs to be updated
          schema:
            type: string
        - name: status
          in: query
          description: Status of object that needs to be updated
          schema:
            type: string
      responses:
        '400':
          description: Invalid input
            
            
    delete:
      tags:
        - Object
      summary: Deletes an object
      description: delete an object
      operationId: deleteObject
      parameters:
        - name: api_key
          in: header
          description: ''
          required: false
          schema:
            type: string
        - name: objectId
          in: path
          description: Object id to delete
          required: true
          schema:
            type: string
      responses:
        '400':
          description: Invalid object value
            
components:
  schemas:
    Object:
      required:
        - name
        - description
      type: object
      properties:
        id:
          type: string
          example: us8a-8dha-aiss-a93h
        name:
          type: string
          example: Object 1
        description:
          type: string
          example: Description of object